+++ Sources/HIGThemesContract/HIGThemeableView.swift
@@ -13,6 +13,7 @@ public struct HIGThemeableView<Content: View>: View {
     public var body: some View {
         content()
             .environment(\.higThemeStorage, theme)
+            .background(HIGThemeRegistrationView(theme: theme))
             .modifier(HIGSizeClassModifier())
     }
 }